style: fold keepalive test into transport_layer_test.go, trim doc comment
- Move TestTransportLayerKeepaliveHandlerTCP into transport_layer_test.go
  next to the sibling readFilter tests (same TransportTCP/net.Pipe/
  readConnection harness); drop the standalone transport_keepalive_test.go.
- Tighten the KeepaliveHandler doc comment.
